Add orchestrator helpers and update tests
